pyproject.toml
[project]
name = "csv-to-bin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow>=11.3.0",
    "tk>=0.1.0",
]
